handle-error finished

+++ src/NonterminalCharacter/AddExp.h
@@ -9,7 +9,7 @@
 
 class AddExp : public Node {
 public:
-    vector<short> op; // 0-+ 1--
+    vector<int> op; // 0-+ 1--
     void addOp(int op);
 
     AddExp(NodeType nodeType1, int lineNum);
