+++ src/NonterminalCharacter/EqExp.h
@@ -8,7 +8,7 @@
 #include <vector>
 class EqExp : public Node{
 public:
-    vector<short> op; // 0-EQL 1-NEQ
+    vector<int> op; // 0-EQL 1-NEQ
     void addOp(int op);
 
     EqExp(NodeType nodeType1, int lineNum);
