+++ src/NonterminalCharacter/FuncType.h
@@ -8,9 +8,9 @@
 #include <string>
 class FuncType : public Node {
 public:
-    short type = 0; // 0 is int 1 is void
+    int type = 0; // 0 is int 1 is void
     FuncType(NodeType nodeType1, int lineNum);
-    short getType() override ;
+    int getType() override ;
 
 };
