+++ src/NonterminalCharacter/Stmt.h
@@ -12,10 +12,10 @@
 
 class Stmt : public Node {
 public:
-    short type = 0; // 0 normal 1-if 2-for 3-break 4-continue 5-return 6-getint 7-printf
+    int type = 0; // 0 normal 1-if 2-for 3-break 4-continue 5-return 6-getint 7-printf
     string str = "";
     Stmt(NodeType nodeType1, int lineNum);
-    short getType() override;
+    int getType() override;
     void setStr(string str);
     string getStr() override;
