+++ src/NonterminalCharacter/UnaryOp.h
@@ -8,7 +8,7 @@
 
 class UnaryOp : public Node {
 public:
-    short type = 0; // 0-+ 1-- 1-!
+    int type = 0; // 0-+ 1-- 1-!
     UnaryOp(NodeType nodeType1, int lineNum);
 
 };
