+++ src/NonterminalCharacter/VarDef.h
@@ -12,11 +12,11 @@ public:
     explicit VarDef(NodeType nodeType1, int lineNum);
     string word;
     bool Equal = false;
-    short type = 0;
+    int type = 0;
     void setWord(const string& word);
     string getWord() override ;
     bool hasEqual() override ;
-    short getType() override ;
+    int getType() override ;
 
 };
